fix(C-1): fix type truncation in add_sample — Boltzmann weight int-truncated to 0, producing NaN thermodynamics
State::count was int; add_sample took int multiplicity. When
get_global_ensemble() passed Boltzmann weights (doubles in 0.0–1.0)
as the multiplicity argument, they were silently truncated to 0.
log(0) = -inf then poisoned every thermodynamic quantity returned
by the public API (free_energy, entropy, heat_capacity all NaN).

Fix:
- State::count: int → double
- add_sample(double, int=1) → add_sample(double, double=1.0)
- serialize_multiplicities() return type: vector<int> → vector<double>
- Drop now-redundant static_cast<double> in compute() Eigen block

The get_global_ensemble() call site already passes double weights
and now works correctly.

Regression test FractionalMultiplicityNoNaN verifies that
add_sample(-10.0, 0.5) produces finite, non-NaN thermodynamics.

+++ LIB/statmech.cpp
@@ -56,7 +56,7 @@ StatMechEngine::StatMechEngine(double temperature_K)
 
 // ─── add_sample ──────────────────────────────────────────────────────────────
 
-void StatMechEngine::add_sample(double energy, int multiplicity) {
+void StatMechEngine::add_sample(double energy, double multiplicity) {
     ensemble_.push_back({energy, multiplicity});
 }
 
@@ -84,7 +84,7 @@ Thermodynamics StatMechEngine::compute() const {
         Eigen::ArrayXd counts(static_cast<Eigen::Index>(N));
         Eigen::ArrayXd energies(static_cast<Eigen::Index>(N));
         for (std::size_t i = 0; i < N; ++i) {
-            counts(static_cast<Eigen::Index>(i))   = static_cast<double>(ensemble_[i].count);
+            counts(static_cast<Eigen::Index>(i))   = ensemble_[i].count;
             energies(static_cast<Eigen::Index>(i)) = ensemble_[i].energy;
         }
         Eigen::ArrayXd lw = counts.log() - beta_ * energies;
@@ -436,8 +436,8 @@ std::vector<double> StatMechEngine::serialize_energies() const {
     return out;
 }
 
-std::vector<int> StatMechEngine::serialize_multiplicities() const {
-    std::vector<int> out(ensemble_.size());
+std::vector<double> StatMechEngine::serialize_multiplicities() const {
+    std::vector<double> out(ensemble_.size());
     for (size_t i = 0; i < ensemble_.size(); ++i)
         out[i] = ensemble_[i].count;
     return out;
